requirements.txt
aniso8601==7.0.0
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
attrs==19.3.0
autopep8==1.5.6
awscli==1.18.69
awsebcli==3.20.2
backcall==0.1.0
bcrypt==3.2.0
bleach==3.1.1
blessed==1.19.0
blinker==1.4
botocore==1.21.65
Brlapi==0.7.0
cached-property==1.5.1
cement==2.8.2
certifi==2019.11.28
cffi==1.15.0
chardet==3.0.4
chrome-gnome-shell==0.0.0
cliapp==1.20180812.1
Click==7.0
cmdtest==0.32+git
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
defusedxml==0.6.0
distlib==0.3.1
distro==1.4.0
Django==3.0.6
django-cors-headers==3.5.0
django-environ==0.4.5
django-phonenumber-field==5.0.0
django-rest-passwordreset==1.1.0
django-seed==0.2.2
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
docker==4.1.0
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
docutils==0.16
entrypoints==0.3
Faker==4.1.2
filelock==3.0.12
future==0.16.0
graphene==2.1.8
graphene-django==2.13.0
graphql-core==2.3.2
graphql-relay==2.0.1
hidpidaemon==18.4.6
html5lib==1.0.1
httplib2==0.14.0
idna==2.8
importlib-metadata==1.5.0
install==1.3.3
ipykernel==5.2.0
ipython==7.13.0
ipython_genutils==0.2.0
ipywidgets==6.0.0
isort==5.4.2
jedi==0.15.2
Jinja2==2.10.1
jmespath==0.10.0
jsonschema==3.2.0
jupyter-client==6.1.2
jupyter-console==6.0.0
jupyter-core==4.6.3
kernelstub==3.1.4
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
louis==3.12.0
macaroonbakery==1.3.1
Markdown==3.1.1
MarkupSafe==1.1.0
mccabe==0.6.1
mistune==0.8.4
more-itertools==4.2.0
mysql-client==0.0.1
nbconvert==5.6.1
nbformat==5.0.4
netifaces==0.10.4
notebook==6.0.3
oauthlib==3.1.0
olefile==0.46
packaging==20.3
pandocfilters==1.4.2
paramiko==2.8.0
parso==0.5.2
pathspec==0.5.9
pexpect==4.6.0
phonenumbers==8.12.8
pickleshare==0.7.5
Pillow==7.2.0
pop-transition==1.1.2
prometheus-client==0.7.1
promise==2.3
prompt-toolkit==2.0.10
protobuf==3.6.1
pyasn1==0.4.2
pycairo==1.16.2
pycodestyle==2.7.0
pycparser==2.20
pycups==1.9.73
pydbus==0.6.0
Pygments==2.3.1
PyGObject==3.36.0
PyJWT==1.7.1
pylint==2.6.0
pymacaroons==0.13.0
PyMySQL==0.10.0
PyNaCl==1.3.0
pyparsing==2.4.6
pyRFC3339==1.1
pyrsistent==0.15.5
python-apt===2.1.2pop0-1587756471-20.04-cd2988e
python-dateutil==2.8.1
python-debian===0.1.36ubuntu1
python-xlib==0.23
pytz==2020.1
pyxdg==0.26
PyYAML==5.3.1
pyzmq==18.1.1
qrcode==6.1
redis==4.0.0b3
repoman==1.2.2
requests==2.24.0
requests-unixsocket==0.2.0
roman==2.0.0
rsa==4.0
Rx==1.6.1
s3transfer==0.3.3
SecretStorage==2.3.1
semantic-version==2.8.5
Send2Trash==1.5.0
sessioninstaller==0.0.0
simplejson==3.16.0
singledispatch==3.4.0.3
six==1.14.0
sqlparse==0.3.1
systemd-python==234
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.4
text-unidecode==1.3
texttable==1.6.2
toml==0.10.1
tornado==5.1.1
traitlets==4.3.3
ttystatus==0.38
ubuntu-advantage-tools==27.2
ubuntu-drivers-common==0.0.0
ufw==0.36
Unidecode==1.1.1
urllib3==1.26.7
virtualenv==20.0.17
virtualenv-clone==0.5.4
wadllib==1.3.3
wcwidth==0.1.8
webencodings==0.5.1
websocket-client==0.53.0
widgetsnbextension==2.0.0
wrapt==1.12.1
xkit==0.0.0
xlrd==1.2.0
xlrd2==1.2.6
xlwt==1.3.0
zipp==1.0.0
